* Move cron from retract_pids file to main * Make generate_cron_yaml.sh OSX friendly
